Check operand type on explicit type Rename TypeId to TypeName Change implicit coercion

+++ src/main/java/org/apache/hop/expression/operator/BitCountFunction.java
@@ -31,7 +31,7 @@ public class BitCountFunction extends Function {
     super(
         "BIT_COUNT",
         ReturnTypes.INTEGER_NULLABLE,
-        OperandTypes.NUMERIC,
+        OperandTypes.INTEGER,
         OperatorCategory.BITWISE,
         "/docs/bit_count.html");
   }
